+++ src/app/projects/add-edit-project/add-edit-project.component.scss
@@ -4,3 +4,13 @@
   border-bottom: 1px solid lightgray;
   margin-bottom: 2rem;
 }
+
+// .nav-tabs {
+//     background: none;
+// }
+// .project-tabs {
+// background: #f7f8fa;
+// }
+// .nav-tabs .nav-link.active {
+// background: #f7f8fa;
+// }
