+++ src/assets/styles/components/tab-nav.scss
@@ -1,67 +1,139 @@
-$nav-tab-border-width: 0;
-$nav-tab-border-color: #fff;
-$nav-tab-bg: #f7f8fa;
-$nav-tab-pane-bg: #fff;
-$nav-tab-bg-active: #fff;
+.nav-tabs {
+    @include flex-flow(row wrap);
+    border-bottom: none;
+
+    .nav-link {
+        padding: 0.75rem 0.2rem;
+        border: none;
+        background: $tab-sm-bg;
+        text-align: center;
+        text-transform: uppercase;
+        font-weight: 700;
+        transition: all ease-in 0.2s;
+    }
+}
+
+// @media (max-width: 767px) {
+//     .nav-tabs {
+//         margin: -0.15rem;
+
+//         .nav-link {
+//             margin: 0.15rem;
+//             border-radius: 0;
+
+//             &.active {
+//                 color: $tab-sm-active-color;
+//                 background-color: $tab-sm-active-bg;
+//             }
+//         }
+//     }
+// }
+
 
 .nav-tabs {
-  border-color: $nav-tab-border-color;
-  border-width: $nav-tab-border-width;
-  background: $nav-tab-bg;
-  text-transform: normal;
+    @include flex-flow(row nowrap);
+    margin-bottom: 2.5rem;
+    border-bottom: 2px solid $tab-nav-border-color;
+    background-color: $tab-nav-bg;
 
-  .nav-item {
-    margin-bottom: calc(#{$nav-tab-border-width} * -1);
-  }
-
-  .nav-link {
-    min-width: 6rem;
-    padding: 1rem 1.5rem;
-    color: #999;
-    border-width: $nav-tab-border-width;
-    border-bottom-width: 0;
-    border-radius: 0;
-    text-align: center;
-    font-size: 0.9375rem;
-    font-weight: 700;
-
-    &:focus,
-    &:hover {
-      color: #333;
-      border: $nav-tab-border-width solid transparent;
-      border-bottom-width: 0;
-      background: rgba($nav-tab-bg,0.5);
+    .nav-item {
+        margin-bottom: -2px;
+        display: inline-block;
+        width: auto;
     }
 
-    &.active {
-      color: #000;
-      border-color: $nav-tab-border-color;
-      border-width: $nav-tab-border-width;
-      border-bottom-width: 0;
-      background: $nav-tab-bg-active;
+    .nav-link {
+        position: relative;
+        padding: 0.75rem;
+        color: rgba($tab-color, 0.8);
+        border-bottom-left-radius: 0;
+        border-bottom-right-radius: 0;
+        background-color: transparent;
+        font-size: 0.85rem;
 
-      &:focus,
-      &:hover {
-        border-bottom-color: transparent;
-        border: $nav-tab-border-width solid $nav-tab-border-color;
-        border-bottom-width: 0;
-      }
+        &:hover {
+            background-color: $tab-hover-bg;
+            box-shadow: 0 -2px 0 $tab-nav-border-color inset;
+        }
+
+        &.active {
+            color: $tab-color;
+            background-color: $tab-active-bg;
+            box-shadow: 0 2px 0 $tab-nav-border-color inset, 2px 0 $tab-nav-border-color inset, -2px 0 $tab-nav-border-color inset;
+        }
     }
-
-    i {
-      margin-top: -0.2rem;
-      margin-right: 0.20rem;
-      margin-left: -0.2rem;
-      vertical-align: middle;
-    }
-  }
 }
 
-.mat-tab-group {
-  font-family: 'Noto Sans', sans-serif;
-}
 
-.tab-pane {
-  padding: 2rem 1.5rem;
-  background: $nav-tab-pane-bg;
-}
+
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+
+
+// $nav-tab-border-width: 0;
+// $nav-tab-border-color: #fff;
+// $nav-tab-bg: #f7f8fa;
+// $nav-tab-pane-bg: #fff;
+// $nav-tab-bg-active: #fff;
+
+// .nav-tabs {
+//   border-color: $nav-tab-border-color;
+//   border-width: $nav-tab-border-width;
+//   background: $nav-tab-bg;
+//   text-transform: normal;
+
+//   .nav-item {
+//     margin-bottom: calc(#{$nav-tab-border-width} * -1);
+//   }
+
+//   .nav-link {
+//     min-width: 6rem;
+//     padding: 1rem 1.5rem;
+//     color: #999;
+//     border-width: $nav-tab-border-width;
+//     border-bottom-width: 0;
+//     border-radius: 0;
+//     text-align: center;
+//     font-size: 0.9375rem;
+//     font-weight: 700;
+
+//     &:focus,
+//     &:hover {
+//       color: #333;
+//       border: $nav-tab-border-width solid transparent;
+//       border-bottom-width: 0;
+//       background: rgba($nav-tab-bg,0.5);
+//     }
+
+//     &.active {
+//       color: #000;
+//       border-color: $nav-tab-border-color;
+//       border-width: $nav-tab-border-width;
+//       border-bottom-width: 0;
+//       background: $nav-tab-bg-active;
+
+//       &:focus,
+//       &:hover {
+//         border-bottom-color: transparent;
+//         border: $nav-tab-border-width solid $nav-tab-border-color;
+//         border-bottom-width: 0;
+//       }
+//     }
+
+//     i {
+//       margin-top: -0.2rem;
+//       margin-right: 0.20rem;
+//       margin-left: -0.2rem;
+//       vertical-align: middle;
+//     }
+//   }
+// }
+
+// .mat-tab-group {
+//   font-family: 'Noto Sans', sans-serif;
+// }
+
+// .tab-pane {
+//   padding: 2rem 1.5rem;
+//   background: $nav-tab-pane-bg;
+// }
